Several code fixes
